- Remote flow controller can merge small writes into a single DATA frame
- Add listener to remote flow-controller that notifies the number of bytes written for a stream.
- Local flow-controller defers writing WINDOW_UPDATE until asked to do so, this allows for a single WINDOW_UPDATE to be written per stream in a read-write cycle as opposed to one update each time the number of returned bytes triggers the threshold.

This will improve throughput in a number of ways
- Coalescing payloads into a single DATA frame causes fewer onDataRead calls on the remote side
- Applications can use the flow-controller listener to receive more granular notifications of written bytes rather than listening to each individual payload
- Deferring WINDOW_UPDATE allows more flow-control to be returned to the remote side as all consumed bytes are returned and not just the amount consumed when the threshold was crossed.

Fixes
https://github.com/netty/netty/issues/3852