@nmittler 2015-06-04
@louiscryan just a few minor comments.  Looks great!