@Scottmitch 2015-06-04
@louiscryan - Great work! Do you have any metrics to report for this change? You mention about [2x](https://github.com/netty/netty/issues/3852#issuecomment-109011983) improvement, but can you share some benchmarks?

When you squash can you just fix up the comment to have the usual 3 sections?