@louiscryan 2015-06-04
Sure do. From this benchmark in GRPC

https://github.com/grpc/grpc-java/blob/master/benchmarks/src/jmh/java/io/grpc/benchmarks/netty/FlowControlledMessagesPerSecondBenchmark.java

I see 445906.159 messages/sec before and 715513.721 messages/sec after.  Not quite 2x but I haven't done anything in GRPC to take advantage of the batchyness yet either

AFTER

Benchmark                                                          (channelCount)  (clientExecutor)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt       Score       Error  Units

FlowControlledMessagesPerSecondBenchmark.stream                                 1            DIRECT                       1           SMALL  thrpt   20       0.999 ±     0.001  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   20  715513.721 ± 48191.126  ops/s


BEFORE

Benchmark                                                          (channelCount)  (clientExecutor)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt       Score       Error  Units
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   20  445906.159 ± 22915.786  ops/s