@Scottmitch 2015-06-04
@louiscryan - Just glanced at the benchmarks...I see a lot of sleeps...interesting setup with the counters. Any thoughts on running with other payload sizes?