@louiscryan 2015-06-04
Take a look at StreamingResponseBandwidthBenchmark which is more focused on measuring bandwidth in bytes/sec instead of messages/sec but uses the same flow-control feedback mechanisms under the hood