@Scottmitch 2015-06-04
@louiscryan - Have you run either of these benchmarks with `MEDIUM`, `LARGE`, `JUMBO` sizes?