@louiscryan 2015-06-05
Bandwidth seems to be down. Not clear yet if this is because we haven't adapted to the change properly in GRPC or not yet. Digging in ....

Before

Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt     Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt   20  3031.875 ±  98.287  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           JUMBO  thrpt   20  2684.513 ±  93.582  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt   20  7148.258 ± 484.105  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           JUMBO  thrpt   20  4336.411 ± 465.072  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt   20  3195.302 ± 118.296  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           JUMBO  thrpt   20  2722.945 ±  95.696  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt   20  6760.362 ± 314.892  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           JUMBO  thrpt   20  4443.045 ± 497.565  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt   20  3088.441 ±  86.876  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           JUMBO  thrpt   20  2594.376 ± 108.244  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt   20  6660.088 ± 213.953  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           JUMBO  thrpt   20  4407.278 ± 386.549  ops/s



After

Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt     Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt   20  2269.320 ± 191.571  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           JUMBO  thrpt   20  2607.399 ± 141.197  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt   20  5115.398 ± 225.719  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           JUMBO  thrpt   20  4214.503 ± 187.410  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt   20  2246.197 ± 171.291  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           JUMBO  thrpt   20  2710.631 ± 127.857  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt   20  4904.255 ± 166.462  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           JUMBO  thrpt   20  4135.827 ± 207.884  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt   20  2384.662 ± 244.015  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           JUMBO  thrpt   20  2508.755 ± 120.795  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt   20  5224.336 ± 320.337  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           JUMBO  thrpt   20  4065.964 ± 185.830  ops/s