@louiscryan 2015-06-06
OK. The reduction in bandwidth for that benchmark is caused by the delay to the write of of the window_update which I was doing after onReadComplete.

If I more aggressively flush window updates I now get slightly better bandwidth than before which is what I was expecting.

e.g.

StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond LARGE 1 LARGE thrpt 20 3406.499 ± 128.233 ops/s

While playing with the local flow-controller I had a thought that I wanted to run by you, @nmittler & @ejona86. Currently we use the a ratio of each individual streams consumption to trigger a window update write. Rather than doing this it might be more sensible to see if the consumption for the connection has exceed a ratio and then return all consumed bytes. i.e. treat the refresh trigger as a function of the connection (and therefore the banwidth delay product) rather than a function of each individual stream.