@louiscryan 2015-06-11
Before

```
Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt     Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt   20  3304.626 ± 110.943  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           JUMBO  thrpt   20  2777.045 ± 124.364  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt   20  2774.740 ±  99.287  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           JUMBO  thrpt   20  2096.918 ± 495.754  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt   20  6993.063 ± 433.202  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           JUMBO  thrpt   20  5193.056 ± 122.303  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt   20  8265.015 ± 417.371  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           JUMBO  thrpt   20  4996.989 ± 177.047  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt   20  7384.874 ± 490.910  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           JUMBO  thrpt   20  5897.575 ± 286.991  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt   20  8841.500 ± 710.635  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           JUMBO  thrpt   20  4915.330 ± 357.087  ops/s
```

After

```
Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt     Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt   20  3017.654 ± 187.414  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           JUMBO  thrpt   20  2416.572 ± 107.467  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt   20  2171.331 ±  95.648  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           JUMBO  thrpt   20  1633.122 ± 390.833  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt   20  6889.882 ± 430.730  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           JUMBO  thrpt   20  5632.200 ± 185.352  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt   20  7958.968 ± 281.252  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           JUMBO  thrpt   20  4980.166 ± 483.920  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt   20  7232.370 ± 427.343  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           JUMBO  thrpt   20  5299.769 ± 461.990  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt   20  9088.211 ± 674.705  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           JUMBO  thrpt   20  4566.509 ± 417.478  ops/s
```