@louiscryan 2015-06-11
Here's the numbers from the other benchmark with some payload size variety.

Before
```
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   20  421393.808 ± 29778.256  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1          MEDIUM  thrpt   20  203715.454 ±  7310.145  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           LARGE  thrpt   20    5839.479 ±   225.818  ops/s
```
After
```
Benchmark                                                          (channelCount)  (clientExecutor)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt       Score       Error  Units
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           SMALL  thrpt   20  711373.623 ± 34085.676  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1          MEDIUM  thrpt   20  261626.277 ± 13361.214  ops/s
FlowControlledMessagesPerSecondBenchmark.stream:messagesPerSecond               1            DIRECT                       1           LARGE  thrpt   20    5486.760 ±   210.770  ops/s
```