@louiscryan 2015-06-17
Update benchmarks. FWIW I have 3 runs, the first two don't include this feature and are just GRPC at head using either 4.1Beta5 (release) or 4.1Beta6 (4.1 branch at head). The last run is with the change patched into the 4.1 branch at head and a modified version of GRPC to use the new APIs.

Either there is some slowdown between 4.1Beta5 and 4.1 HEAD that is not correlated with this PR or there is enough noise in my benchmarking environment that my runs are a little unstable inherently. Either way max variation across any run set does not exceed 10% which seems acceptable


Before 4.1B5

Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt      Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt  100   3319.727 ±  58.141  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt  100   2984.916 ±  49.391  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt  100   7714.442 ± 151.093  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt  100   9730.384 ± 265.793  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt  100   7901.952 ± 126.430  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt  100  10225.277 ± 310.226  ops/s

Before 4.1B6

Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt      Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt  100   3198.594 ±  55.484  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt  100   2722.837 ±  34.891  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt  100   7607.570 ±  98.890  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt  100   9436.205 ± 246.687  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt  100   7948.061 ± 124.580  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt  100  10087.004 ± 279.296  ops/s

After 4.1B6

Benchmark                                                     (clientInboundFlowWindow)  (maxConcurrentStreams)  (responseSize)   Mode  Cnt      Score     Error  Units
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                       1           LARGE  thrpt  100   3065.619 ±  68.655  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                     MEDIUM                      10           LARGE  thrpt  100   2707.967 ±  42.574  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                       1           LARGE  thrpt  100   7565.128 ±  99.997  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      LARGE                      10           LARGE  thrpt  100   8956.562 ± 182.238  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                       1           LARGE  thrpt  100   7743.380 ± 126.945  ops/s
StreamingResponseBandwidthBenchmark.stream:megabitsPerSecond                      JUMBO                      10           LARGE  thrpt  100   9940.198 ± 265.941  ops/s